Support syncing user groups from GitHub
Add hooks that allow Trac to use groups generated from and
organization's GitHub teams for permission management. To keep the
groups in sync, add an endpoint for a GitHub WebHook that listens for
membership events and updates the cache.

Optionally support signed webhooks using a shared secret to prevent
denial of service attacks.

Synced groups will have the form
 - github-$org-$teamslug for members of a certain group, and
 - github-$org for the union of all members of all teams in this org.

Note that the github-$org group does not contain users that are part of
the GitHub organization, but no team inside that organization. This is
deliberate because GitHub currently does not provide a notification
mechanism for changes to organization membership, which means we would
have to resort to polling.

In our intended use case, this difference does not matter. Commented
code implements syncing of GitHub org memberships should GitHub decide
to provide a WebHook in the near future.

Signed-off-by: Clemens Lang <cal@macports.org>