Run trac-github-update hook after updating repo
Git has no option to run a hook locally when 'git remote update' fetches
new commits. In order to be able to trigger custom actions on update
(for example commit mail notifications), this adds suppport for
a 'trac-github-update' hook. This hook will be called each time the
GitHub Webhook updates the repository, and receives the JSON payload as
sent by GitHub on standard input.