Support prefixing usernames for GitHub users
Allow prefixing usernames of users logged in via GitHub (e.g.
gh:<username>) so as to distinguish them from legacy users on the
system.

This option is needed from some of the other components, so to make that
easy I added it to the GitHubMixin and made it a subclass of the login
component.