Fix failures introduced by f859fb8
The change in f859fb8 assumed that GitHub would respond to API requests
to /organizations/$org_name/team/$team_id, but it doesn't. The only
supported URL seems to be /organizations/$org_id/team/$team_id, so we
need to somehow query the teams using both their org and team IDs.
Fortunately GitHub includes an "url" field in the response for listing
teams that can be used to obtain the correct path.

Use this field to determine the correct URL and adjust the test harness
to match GitHub's actual behavior.